refactor(reminder-message): move db logic to central server
