+++ packages/facility-server/app/services/websocketClientService.js
@@ -16,25 +16,12 @@ export const defineWebsocketClientService = injector => {
      * @param {{ contactId: string, chatId: string }} payload
      */
     async ({ chatId, contactId }) => {
-      const contact = await injector.models?.PatientContact.findByPk(contactId, {
-        include: [{ model: injector.models?.Patient, as: 'patient' }],
-      });
-
+      const contact = await injector.models?.PatientContact.findByPk(contactId);
       if (!contact) return;
 
       contact.connectionDetails = { chatId };
       await contact.save();
 
-      const contactName = contact.name;
-      const patientName = [
-        contact.patient.firstName,
-        contact.patient.middleName,
-        contact.patient.lastName,
-      ].join(' ');
-
-      const successMessage = `Dear ${contactName}, you have successfully registered to receive messages for ${patientName}. Thank you.`; //TODO: translate this
-
-      client.emit('telegram:send-message', { chatId, message: successMessage });
       injector.websocketService.emit('telegram:subscribe:success', { contactId, chatId });
     },
   );
